updated huggingface client code
